Rename commit delete
