+++ luact/fiber/commit_delete/init.lua
@@ -27,26 +27,25 @@
 --]]
 local tags = require "luact.tags"
 
-local commit_delete_host = require "luact.fiber.commit_delete.commit_delete_host"
-local commit_delete_with_hooks = require "luact.fiber.commit_delete.commit_delete_with_hooks"
+local with_hooks = require "luact.fiber.commit_delete.with_hooks"
+local with_host = require "luact.fiber.commit_delete.with_host"
 
 local function commit_delete(work_in_progress)
-  -- Iterate all children
-  local child = work_in_progress.child
-  while (child) do
-    commit_delete(work_in_progress.child)
-    child = child.sibling
-  end
-
   -- Perform effect
   if (work_in_progress.type == tags.type.HOST) then
-    commit_delete_host(work_in_progress)
+    with_host(work_in_progress)
   end
   if (
     work_in_progress.type == tags.type.COMPONENT
     or work_in_progress.type == tags.type.MEMO
   ) then
-    commit_delete_with_hooks(work_in_progress)
+    with_hooks(work_in_progress)
+  end
+  -- Iterate all children
+  local child = work_in_progress.child
+  while (child) do
+    commit_delete(child)
+    child = child.sibling
   end
 end
