+++ luact/fiber/commit_delete/with_hooks.lua
